.devdev/streamlit-tasks-dashboard/pyproject.toml
[project]
name = "streamlit-tasks-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.4.0",
    "streamlit>=1.50.0",
]
